WIP: factor out BatchTexture
